+++ lib/iterable.dart
@@ -1,6 +1,7 @@
 /// Some fancy iterables and iterators.
 export 'src/iterable/chunked.dart' show ChunkedExtension;
 export 'src/iterable/combinations.dart' show CombinationsExtension;
+export 'src/iterable/count.dart' show CountIterableExtension;
 export 'src/iterable/flat_map.dart' show FlatMapExtension;
 export 'src/iterable/flatten.dart' show FlattenExtension;
 export 'src/iterable/group.dart' show GroupExtension, Group;